SPS scripts now installed to share directory
By default, installs the processing scripts to

/usr/local/share/spiceweasel/

By using --prefix=/some/directory/

this is changed to

/some/directory/share/spicweasel

and some other configure options can be used to set it as well. See --help.

The directory used is passed to the build (through
defining DEFAULT_SPS_PATH), and then used to search
for a script.